@@ -3,46 +3,51 @@
   "version": "1.0.0",
   "private": true,
   "scripts": {
-    "build": "set NODE_ENV=production&&webpack -p",
+    "build": "set NODE_ENV=production&&webpack -p --progress",
     "webpack": "webpack-dev-server -d --port 8080 --hot --inline --content-base dist/ --history-api-fallback",
     "analyze": "set NODE_ENV=production&&webpack -p --json | webpack-bundle-size-analyzer"
   },
   "devDependencies": {
     "array.prototype.find": "^2.0.0",
     "array.prototype.findindex": "^2.0.0",
-    "babel-core": "6.8.0",
-    "babel-loader": "6.2.4",
-    "babel-plugin-transform-object-assign": "6.8.0",
-    "babel-plugin-transform-object-rest-spread": "6.8.0",
-    "babel-plugin-transform-react-remove-prop-types": "0.2.6",
-    "babel-polyfill": "6.8.0",
-    "babel-preset-es2015": "6.6.0",
-    "babel-preset-react": "6.5.0",
-    "css-loader": "0.23.1",
+    "babel-core": "6.26.3",
+    "babel-loader": "^7.1.5",
+    "babel-plugin-transform-object-assign": "6.22.0",
+    "babel-plugin-transform-object-rest-spread": "6.26.0",
+    "babel-plugin-transform-react-remove-prop-types": "0.4.20",
+    "babel-polyfill": "6.26.0",
+    "babel-preset-env": "^1.7.0",
+    "babel-preset-react": "6.24.1",
+    "create-react-class": "^15.6.3",
+    "css-loader": "1.0.1",
     "dnn-global-styles": "0.0.5",
-    "dnn-webpack-externals": "1.1.2",
+    "dnn-webpack-externals": "1.1.10",
     "es6-object-assign": "^1.0.3",
-    "eslint": "2.9.0",
-    "eslint-loader": "1.3.0",
-    "eslint-plugin-react": "5.1.1",
-    "eslint-plugin-spellcheck": "0.0.6",
-    "file-loader": "0.8.5",
-    "i18n-webpack-plugin": "0.2.7",
-    "jwt-decode": "2.0.1",
-    "less": "2.7.1",
-    "less-loader": "2.2.3",
+    "eslint": "5.9.0",
+    "eslint-loader": "2.1.1",
+    "eslint-plugin-react": "7.11.1",
+    "eslint-plugin-spellcheck": "0.0.11",
+    "file-loader": "2.0.0",
+    "i18n-webpack-plugin": "1.0.0",
+    "jwt-decode": "2.2.0",
+    "less": "3.8.1",
+    "less-loader": "4.1.0",
     "moment": "^2.15.0",
+    "prop-types": "^15.6.2",
     "raw-loader": "0.5.1",
-    "react-click-outside": "^2.1.0",
-    "react-hot-loader": "1.3.0",
-    "style-loader": "0.13.1",
-    "throttle-debounce": "^1.0.1",
-    "url-loader": "0.5.7",
-    "webpack": "1.13.0",
-    "webpack-bundle-size-analyzer": "2.0.1",
-    "webpack-dev-server": "1.14.1"
+    "react": "^16.6.3",
+    "react-click-outside": "^3.0.1",
+    "react-dom": "^16.6.3",
+    "react-hot-loader": "4.3.12",
+    "style-loader": "0.23.1",
+    "throttle-debounce": "^2.0.1",
+    "url-loader": "1.1.2",
+    "webpack": "4.26.0",
+    "webpack-bundle-size-analyzer": "3.0.0",
+    "webpack-cli": "^3.1.2",
+    "webpack-dev-server": "3.1.10"
   },
   "dependencies": {
-    "react-widgets": "^3.4.4"
+    "react-widgets": "^4.4.6"
   }
 }
